+++ scripts/build_site.sh
@@ -11,3 +11,7 @@ node bin/typedoc --json docs/docs.json --readme none
 
 # Build the actual site, references the API docs
 node bin/typedoc --options site/typedoc.config.jsonc
+
+# Create/copy static files
+node scripts/generate_options_schema.js docs-site/schema.json
+cp site/googlede482cdb17c37ad4.html docs-site/googlede482cdb17c37ad4.html